a syntactical suggestion on how to solve the extraction / late execution dilemma
The basic idea is not to deliver the extracted values themselves to subroutes
but instead some token which can then later be used to get the actual values.
This has the advantage that sub-routes can be calculated at route building time
and not only when the actual value is available.

Two added complications:
  - a new directive `extractHere` which demarcates route building and actual extraction time
  - a possible conflation of the Extract token with its value in contexts where
    it is not clear what is meant. In the example we have to use `agent.get` to
    designate that we want to extract the value and don't mean the token. This
    is particularly dangerous in the string building context where you would probably
    never mean the token but the token would always fit type-wise.